# backend/app/admin/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: Kratos Admin API
    description: Kratos Admin API
    contact:
        name: tx7do
        url: https://github.com/tx7do/kratos-admin
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/kratos-admin/blob/master/LICENSE
    version: "1.0"
paths:
...
    /admin/v1/private_messages:
        get:
            tags:
                - PrivateMessageService
            description: 查询私信消息列表
            operationId: PrivateMessageService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  description: 每一页的行数
                  schema:
                    type: integer
                    format: int32
                - name: query
                  in: query
                  description: AND过滤参数，其语法为json格式的字符串，如：{"key1":"val1","key2":"val2"}，具体请参见：https://github.com/tx7do/go-utils/tree/main/entgo/query/README.md
                  schema:
                    type: string
                - name: or
                  in: query
                  description: OR过滤参数，语法同AND过滤参数。
                  schema:
                    type: string
                - name: orderBy
                  in: query
                  description: 排序条件，其语法为JSON字符串，例如：{"val1", "-val2"}。字段名前加'-'为降序，否则为升序。
                  schema:
                    type: array
                    items:
                        type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
                - name: tenantId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListPrivateMessageResponse'
        post:
            tags:
                - PrivateMessageService
            description: 创建私信消息
            operationId: PrivateMessageService_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreatePrivateMessageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/private_messages/{data.id}:
        put:
            tags:
                - PrivateMessageService
            description: 更新私信消息
            operationId: PrivateMessageService_Update
            parameters:
                - name: data.id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdatePrivateMessageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/private_messages/{id}:
        get:
            tags:
                - PrivateMessageService
            description: 查询私信消息详情
            operationId: PrivateMessageService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PrivateMessage'
        delete:
            tags:
                - PrivateMessageService
            description: 删除私信消息
            operationId: PrivateMessageService_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
...
        CreatePrivateMessageRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/PrivateMessage'
            description: 创建私信消息 - 请求
...
        ListPrivateMessageResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/PrivateMessage'
                total:
                    type: integer
                    format: uint32
            description: 查询私信消息列表 - 回应
...
        PrivateMessage:
            type: object
            properties:
                id:
                    type: integer
                    description: 消息ID
                    format: uint32
                subject:
                    type: string
                    description: 主题
                content:
                    type: string
                    description: 内容
                status:
                    enum:
                        - PrivateMessageStatus_Unknown
                        - PrivateMessageStatus_Draft
                        - PrivateMessageStatus_Sent
                        - PrivateMessageStatus_Received
                        - PrivateMessageStatus_Read
                        - PrivateMessageStatus_Archived
                        - PrivateMessageStatus_Deleted
                    type: string
                    description: 消息状态
                    format: enum
                senderId:
                    type: integer
                    description: 发送者用户ID
                    format: uint32
                senderName:
                    type: string
                    description: 发送者用户名称
                senderAvatar:
                    type: string
                    description: 发送者用户头像
                receiverId:
                    type: integer
                    description: 接收者用户ID
                    format: uint32
                receiverName:
                    type: string
                    description: 接收者用户名称
                receiverAvatar:
                    type: string
                    description: 接收者用户头像
                createTime:
                    type: string
                    description: 创建时间
                    format: date-time
                updateTime:
                    type: string
                    description: 更新时间
                    format: date-time
                deleteTime:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 私信消息
...
        UpdatePrivateMessageRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/PrivateMessage'
                updateMask:
                    example: id,realName,userName
                    type: string
                    description: 要更新的字段列表
                    format: field-mask
                allowMissing:
                    type: boolean
                    description: 如果设置为true的时候，资源不存在则会新增(插入)，并且在这种情况下`updateMask`字段将会被忽略。
            description: 更新私信消息 - 请求
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh_token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...